+++ src/image/Image.cc
@@ -516,9 +516,7 @@ int Image::disk_attribute(  VectorAttribute * disk,
     disk->replace("IMAGE",    name);
     disk->replace("IMAGE_ID", iid.str());
     disk->replace("SOURCE",   source);
-
-    // TODO: move all size values (image, volatile disk, quotas) to long long
-    disk->replace("SIZE",     (int)size_mb);
+    disk->replace("SIZE",     size_mb);
 
     if (driver.empty() && !template_driver.empty())//DRIVER in Image,not in DISK
     {
