+++ src/scheduler/include/DatastoreXML.h
@@ -40,7 +40,7 @@ public:
      *    @param vm_disk_mb capacity needed by the VM
      *    @return true if the datastore can host the VM
      */
-    bool test_capacity(unsigned int vm_disk_mb) const
+    bool test_capacity(long long vm_disk_mb) const
     {
         return (vm_disk_mb < free_mb);
     };
@@ -50,7 +50,7 @@ public:
      *    @param vm_disk_mb capacity needed by the VM
      *    @return 0 on success
      */
-    void add_capacity(unsigned int vm_disk_mb)
+    void add_capacity(long long vm_disk_mb)
     {
         free_mb  += vm_disk_mb;
     };
@@ -70,7 +70,7 @@ private:
     int oid;
     int cluster_id;
 
-    unsigned int free_mb; /**< Free disk for VMs (in Mb). */
+    long long free_mb; /**< Free disk for VMs (in MB). */
 
     static const char *ds_paths[]; /**< paths for search function */
